Add basic structure for address mapping
All of the abstraction for the page table construction are prepared now.
What we need to do for the next stage is walk the page table and set the
page table entries correctly.

On the other hand, I'm considering if we could let the access of
page::alloc pointer more easily by some macro. Because we could
convert it to any other type of pointer in the future.