Finish the linear mapping of kernel
The work to map kernel using 4K page resolution is completed.
However, the access permission of different segment is not done.

By the way, the access of the pointer form page::alloc is still
not good. Because of many limitation for Rust to do type
conservation of a pointer, I think the codes are not very easy to
be read.