Map code and stack section for new created task
Since we're going to use another page table for different
tasks, we have to modify several naive implementations which
is applied now.

TODO list:
1. We have to map both kernel and user space to the same virtual
address of trap vector's entry point. It could just be a small
trampoline which helps to swap the page table. Because of the reason,
linear mapping for the kernel space may not be the best idea.
2. Implement interface to swap between kernel and user perfectly
3. After everything is ready, try to spawn userspace task and see if it
run well