Delegate exception to S-mode with syscall API
Please fix:
Don't context switch in the machine timer interrupt handler
directly. Because machine mode interrupt is always allowed to
preempt the supervisor interrupt handler, we may miss the
chance to update the field 'pc' in the TrapFrame in time,
and switch to an incorrect PC at the end.

Instead, just trigger another supervisor interrupt for indirect
context switch when hitting machine timer interrupt. Since we can
control the hardware to avoid nested supervisor interrupt, we can
make sure to switch to the right PC.

TODO:
1. Update the field 'pc' in the TrapFrame for correct context switch
2. Modify timer interrupt handling which causes task preemption