Switch task under S-mode and change related design
Some changes are included in this commit to switch
between userspace and kernel space:

1. Instead of context switching directly in the machine
timer interrupt handler, we just trigger another supervisor
interrupt when timer tick, and context switch for this interrupt
under supervisor mode. This avoid to context switch during
supervisor interrupt service routine which is a complicated scenario
to be handled.
2. Since we decide to context switch at supervisor mode, which has
virtual address translation, the function for switching and the trap
frame have to be mapped for each task
3. To switch to a specific userspace task, we should enable it
to access its own trap frame for context restoring. It means
that we need to map a virtual address for the trap frame.
For the purpose to better manage it in kernel, we assign a
fixed virtual address for trap frame.

However, these changes are still not enough to switch correctly :(