Rewrite trap handler for correct context switch
It is a bad news that we find the orignal design has a lot
of defects to complete the context switch between kernel
and user task :( Because of the reason, the interaction
between timer interrupt and context switch is going to
have a big change.

In this commit, we first modify the way of how the context
being backup if entering interrupt handler. We also stop
the scheduler to pick the next task now, because it could be
more simply to rewrite it from scratch instead of basing on the
problematic implementation.