Finish implementation of mknod()
This commit also change the mechanism to synchronize
on-mem inode to disk, which depends on the object lifetime.
However, such mechanism will have error when we have two
on-mem inode which related to a same inum. Let's leave this
scenario as impossible not but we should take care for this
in the future.